feat: renamed app/controllers/blueprints to routes
